ocr-service/requirements.txt
requests
pillow
opencv-python
pymupdf
numpy
fastapi
uvicorn
python-multipart
python-dotenv
transformers
torch